On i386, produce binaries that are compatible with our W^X implementation.
Note that this will still produce binaries that have the read-only segment
placed before the executable segment.  This means that on machines without
the NX bit, the read-only segment is now executable.  However on machines
with an NX bit, the ordering used by lld means the program headers are no
longer executable since they are now part of the read-only segment.

ok deraadt@, naddy@